# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fashion Search POC API",
    "version": "0.1.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "post": {
        "summary": "Search",
        "operationId": "search_api_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{product_id}": {
      "get": {
        "summary": "Product",
        "operationId": "product_api_v1_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/brands/{slug}": {
      "get": {
        "summary": "Brand",
        "operationId": "brand_api_v1_brands__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionaries": {
      "get": {
        "summary": "Dictionaries",
        "operationId": "dictionaries_api_v1_dictionaries_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/out/{offer_id}": {
      "get": {
        "summary": "Outbound",
        "description": "302 только на URL из базы, прошедший проверку доменов продавца. Произвольный URL не принимается.",
        "operationId": "outbound_api_v1_out__offer_id__get",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/imports": {
      "post": {
        "summary": "Start Import",
        "operationId": "start_import_internal_imports_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/imports/{run_id}": {
      "get": {
        "summary": "Get Import",
        "operationId": "get_import_internal_imports__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/imports/{run_id}/publish": {
      "post": {
        "summary": "Approve Import",
        "operationId": "approve_import_internal_imports__run_id__publish_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FacetValue": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "value",
          "count"
        ],
        "title": "FacetValue"
      },
      "Filters": {
        "properties": {
          "size": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SizeFilter"
              },
              {
                "type": "null"
              }
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "in_stock",
              "any"
            ],
            "title": "Availability",
            "default": "any"
          },
          "price_min_minor": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Min Minor"
          },
          "price_max_minor": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "RUB"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Categories"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Colors"
          },
          "materials": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Materials"
          },
          "brands": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Brands"
          },
          "merchants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Merchants"
          },
          "gender": {
            "items": {
              "type": "string",
              "enum": [
                "women",
                "men",
                "unisex",
                "kids",
                "unknown"
              ]
            },
            "type": "array",
            "title": "Gender"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Filters"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportRequest": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source"
        ],
        "title": "ImportRequest"
      },
      "PublishRequest": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PublishRequest"
      },
      "Relaxation": {
        "properties": {
          "filter": {
            "type": "string",
            "title": "Filter"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "filter",
          "count"
        ],
        "title": "Relaxation"
      },
      "SearchItem": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "brand_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Slug"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "matched_offer_id": {
            "type": "string",
            "title": "Matched Offer Id"
          },
          "price_minor": {
            "type": "integer",
            "title": "Price Minor"
          },
          "old_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Price Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "merchant": {
            "type": "string",
            "title": "Merchant"
          },
          "merchant_slug": {
            "type": "string",
            "title": "Merchant Slug"
          },
          "availability": {
            "type": "string",
            "title": "Availability"
          },
          "available_sizes": {
            "items": {
              "$ref": "#/components/schemas/SizeOut"
            },
            "type": "array",
            "title": "Available Sizes"
          },
          "freshness": {
            "type": "string",
            "enum": [
              "fresh",
              "stale"
            ],
            "title": "Freshness"
          },
          "outbound_url": {
            "type": "string",
            "title": "Outbound Url"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "title",
          "brand",
          "brand_slug",
          "category",
          "color",
          "image_url",
          "matched_offer_id",
          "price_minor",
          "old_price_minor",
          "currency",
          "merchant",
          "merchant_slug",
          "availability",
          "available_sizes",
          "freshness",
          "outbound_url",
          "is_demo",
          "score"
        ],
        "title": "SearchItem"
      },
      "SearchRequest": {
        "properties": {
          "q": {
            "type": "string",
            "maxLength": 500,
            "title": "Q",
            "default": ""
          },
          "filters": {
            "$ref": "#/components/schemas/Filters"
          },
          "disabled_inferred_filters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Disabled Inferred Filters"
          },
          "sort": {
            "type": "string",
            "enum": [
              "relevance",
              "price_asc",
              "price_desc"
            ],
            "title": "Sort",
            "default": "relevance"
          },
          "limit": {
            "type": "integer",
            "maximum": 48.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 24
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "anonymous_session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Anonymous Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "search_id": {
            "type": "string",
            "title": "Search Id"
          },
          "experiment_arm": {
            "type": "string",
            "title": "Experiment Arm"
          },
          "parsed": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parsed"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "$ref": "#/components/schemas/Total"
          },
          "window_limit": {
            "type": "integer",
            "title": "Window Limit"
          },
          "facets": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/FacetValue"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Facets"
          },
          "relaxations": {
            "items": {
              "$ref": "#/components/schemas/Relaxation"
            },
            "type": "array",
            "title": "Relaxations"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "degraded_mode": {
            "type": "boolean",
            "title": "Degraded Mode"
          },
          "ranking_version": {
            "type": "string",
            "title": "Ranking Version"
          },
          "demo_data": {
            "type": "boolean",
            "title": "Demo Data"
          }
        },
        "type": "object",
        "required": [
          "search_id",
          "experiment_arm",
          "parsed",
          "items",
          "total",
          "window_limit",
          "facets",
          "relaxations",
          "next_cursor",
          "degraded_mode",
          "ranking_version",
          "demo_data"
        ],
        "title": "SearchResponse"
      },
      "SizeFilter": {
        "properties": {
          "system": {
            "type": "string",
            "enum": [
              "RU",
              "EU",
              "US",
              "UK",
              "INT",
              "ONE",
              "UNKNOWN"
            ],
            "title": "System"
          },
          "label": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "system",
          "label"
        ],
        "title": "SizeFilter",
        "description": "Размер всегда передаётся вместе с системой: «38» без системы не принимается."
      },
      "SizeOut": {
        "properties": {
          "system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "system",
          "label"
        ],
        "title": "SizeOut"
      },
      "Total": {
        "properties": {
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "relation": {
            "type": "string",
            "enum": [
              "eq",
              "gte"
            ],
            "title": "Relation"
          }
        },
        "type": "object",
        "required": [
          "value",
          "relation"
        ],
        "title": "Total"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}